pickle error pages to file system rather than store internally
I was actually quite fond of the way errors were stored in an internal
dict, served from there and deallocated. However, this fails entirely
when developing in a multithreaded environment (which isn't a must but
is useful if you're doing stuff like AJAX-controlled uploads, etc).

So this patch does away with the internal dict and pickles the error
response objects in a temporary directory.

If the temporary directory does not exist it will be created with mode
1777. Pickles are removed when they are served (ha!). Due to numerous
complications and lazyness on my part, the temporary directory will stay
indefinitely.

Pickles will be ignored if not created by the current user (this is a
somewhat paranoid defence against a borderline ridiculous attack where an
attacker compromises your account by crafting a malicious pickle file
and directing your development server there; in all likelihood, such a
priviliged attacker already owns your arse).